+++ packages/styled-components/src/sheet/dom.js
@@ -21,9 +21,11 @@ const findLastStyleTag = (target: HTMLElement): void | HTMLStyleElement => {
 
 /** Create a style element inside `target` or <head> after the last */
 export const makeStyleTag = (target?: HTMLElement): HTMLStyleElement => {
-  const head = ((document.head: any): HTMLElement);
+  let targetDocument = document;
+  if(target) targetDocument = target.ownerDocument;
+  const head = ((targetDocument.head: any): HTMLElement);
   const parent = target || head;
-  const style = document.createElement('style');
+  const style = targetDocument.createElement('style');
   const prevStyle = findLastStyleTag(parent);
   const nextSibling = prevStyle !== undefined ? prevStyle.nextSibling : null;